Prompt user to change old passwords
When a user navigates to a site for which a password is saved, and the
add-on determines that this password is old, the user is prompted to
change it. The user can choose to change it, be prompted later, or never
be prompted again. (If the user elects to change the password, nothing
happens, for now.)